real_time_simulation_visualizer/requirements.txt
numpy==1.26.4
carla==0.9.15
notebook
pandas
scikit-learn
matplotlib
plotly
tf2onnx
tqdm
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.4.0+cu118
torchvision==0.19.0+cu118
torchaudio==2.4.0+cu118
wheel
setuptools
opencv-python
pyquaternion
av2
kornia==0.6.5
spconv-cu118
dash
dash_canvas
dash-extensions
websockets
flask_sockets
flask-socketio
eventlet
python-socketio
kaleido